% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{bw.CV}
\alias{bw.CV}
\title{Bandwidth Selectors for Kernel Density Estimation}
\usage{
bw.CV(
  x,
  y = NULL,
  kernel = "gaussian",
  robust.iterations = 0,
  degree = 0,
  start.bw = NULL,
  same = FALSE,
  CV = c("DCV", "LSCV"),
  opt.fun = c("nlm", "optim", "nlminb", "optimise"),
  ret.fun = NULL,
  par.name.in.opt = NULL,
  fun.name.in.opt = NULL,
  report.code = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector or numeric matrix.}

\item{y}{A numeric vector of responses (dependent variable) if \code{CV = "LSCV"}.}

\item{kernel}{Passed to \code{kernelDensity} or to \code{kernelSmooth}.}

\item{robust.iterations}{Passed to \code{kernelSmooth} if \code{CV = "LSCV"}.}

\item{degree}{Passed to \code{kernelSmooth} if \code{CV = "LSCV"}.}

\item{start.bw}{Initial value for bandwidth search.}

\item{same}{Logical: use the same bandwidth for all columns of \code{x}?}

\item{CV}{"DCV" for density cross-validation, "LSCV" for least-squares cross-validation.}

\item{opt.fun}{Optimiser name (can be a custom function).}

\item{ret.fun}{A function that extract the optimal parameter from the output of the optimiser.}

\item{par.name.in.opt}{The name of the initial parameter vector in the output of the optimiser.}

\item{fun.name.in.opt}{The name of the criterion in the output of the optimiser (to be minimised).}

\item{report.code}{Logical: print out the optimiser return code for diagnostics?}

\item{...}{Other parameters passed to the optimiser.}
}
\value{
Numeric vector or scalar of the optimal bandwidth.
}
\description{
Finds the optimal bandwidth by minimising the density cross-valication or least-squares criteria.
Remember that since usually, the CV function is highly non-linear, the return value should be taken with a grain of salt.
With non-smooth kernels (such as uniform), it will oftern return the local minimum after starting from a reasonable value.
The user might want to standardise the input matrix \code{x} by column (divide by some estimator of scale, like \code{sd}
or \code{IQR}) and examine the behaviour of the CV criterion as a function of unique bandwidth (\code{same} argument).
If it seems that the optimum is unique, then they may proceed by multiplying the bandwidth by the scale measure,
and start the search for the optimal bandwidth in multiple dimensions.
}
