% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{DCV}
\alias{DCV}
\title{Density cross-validation}
\usage{
DCV(x, bw, same = FALSE, kernel = "gaussian", order = 2)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{bw}{Candidate bandwidth values: scalar, vector, or a matrix (with columns corresponding to columns of \code{x}).}

\item{same}{Logical: use the same bandwidth for all columns of \code{x}?}

\item{kernel}{Passed to \code{kernelWeights}.}

\item{order}{Passed to \code{kernelWeights}.}
}
\value{
A numeric vector of the same length as \code{bw} or \code{nrow(bw)}.
}
\description{
Density cross-validation
}
