% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{decFac}
\alias{decFac}
\title{Bandwidth decrement rule.}
\usage{
decFac(pow)
}
\arguments{
\item{pow}{Convergence power (2 for regular estimators, 5 for non-parametric ones).}
}
\value{
A numeric vector of length 5.
}
\description{
Bandwidth decrement rule.
}
