% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{.fail}
\alias{.fail}
\title{Fail gracefully}
\usage{
.fail()
}
\value{
A list of the same format as the output of \code{\link{maximiseSEL}}.
}
\description{
Return parseable output in case the SEL optimiser fails
}
