% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{.lead0}
\alias{.lead0}
\title{Adding leading or trailing zeros}
\usage{
.lead0(x, d)
}
\arguments{
\item{x}{A number or a numeric vector to format.}

\item{d}{The desired length of the string.}
}
\value{
A character vector.
}
\description{
Adding leading or trailing zeros
}
