% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scel.R
\name{emplik}
\alias{emplik}
\title{Self-concordant empirical likelihood for a vector mean}
\source{
This code was written and published on his personal web site by Art B. Owen (February 2014),
  and slightly reworked by Andreï V. Kostyrka.
}
\usage{
emplik(
  z,
  mu,
  lam,
  eps,
  M,
  thresh = 1e-30,
  itermax = 100,
  ALPHA = 0.3,
  BETA = 0.8,
  BACKEPS = 0,
  verbose = FALSE
)
}
\arguments{
\item{z}{A numeric vector or a matrix with one data vector per column.}

\item{mu}{Hypothesised mean, default (0 ... 0) in R^ncol(z)}

\item{lam}{Starting lambda, default (0 ... 0)}

\item{eps}{Lower cut-off for \code{mllog()}, default 1/nrow(z)}

\item{M}{Upper cutoff for \code{mllog()}, default Inf}

\item{thresh}{Convergence threshold for log-likelihood (the default is aggressive)}

\item{itermax}{Upper bound on number of Newton steps (seems ample)}

\item{ALPHA}{Backtracking line search parameter: acceptance of a decrease in function value by ALPHA*f of the prediction
based on the linear extrapolation.}

\item{BETA}{Backtracking line search reduction factor. 0.1 corresponds to a very crude search, 0.8 corresponds
to a less crude search.}

\item{BACKEPS}{Backtrack threshold: the search can miss by this much. Consider setting it to 1e-10
if backtracking seems to be failing due to round-off.}

\item{verbose}{Logical: print output diagnostics?}
}
\value{
A list with the following values:
\describe{
    \item{logelr}{Log of empirical likelihood ratio (equal to 0 if the hypothesised mean is equal to the sample mean)}
    \item{lam}{Vector of Lagrange multipliers}
    \item{wts}{Observation weights/probabilities (vector of length n)}
    \item{converged}{\code{TRUE} if algorithm converged. \code{FALSE} usually means that mu is not in the convex hull of the data. Then, a very small likelihood is returned (instead of zero).}
    \item{iter}{Number of iterations taken.}
    \item{ndec}{Newton decrement (see Boyd & Vandenberghe).}
    \item{gradnorm}{Norm of the gradient of log empirical likelihood.}
    }
}
\description{
Implements the empirical likelihood test for whether the mean of coordinates of \code{z} is \code{mu}.
}
\details{
Tweak \code{ALPHA} and \code{BETA} with extreme caution. See See Boyd and Vandenberghe, pp. 464--466 for details.
It is necessary that \code{0 < ALPHA < 1/2} and \code{0 < BETA < 1}.
\code{ALPHA = 0.03} seems better than 0.01 on some 2-dimensional test data (sometimes fewer iterations).
}
\examples{
earth <- c(5.5,5.61,4.88,5.07,5.26,5.55,5.36,5.29,5.58,5.65,5.57,5.53,5.62,5.29,
  5.44,5.34,5.79,5.1,5.27,5.39,5.42,5.47,5.63,5.34,5.46,5.3,5.75,5.68,5.85)
emplik(earth, mu = 5.517) # 5.517 is the modern accepted value

}
