% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{estimateOneModelDesign1}
\alias{estimateOneModelDesign1}
\title{Estimation of one model via SEL}
\usage{
estimateOneModelDesign1(
  start.mod,
  data,
  sel.bw,
  eff.control = list(),
  do.SE = FALSE,
  do.restr = FALSE,
  CI.lev = NULL,
  try.ellipse = TRUE,
  ...
)
}
\arguments{
\item{start.mod}{A list that must contain the following elements: \code{coefficients}
(initial values), \code{vcov} (a plausible guess about the variance-covariance of the estimator;
used when teremining the optimiser step), and \code{residuals} (true residuals
for the validation sample, i.e. subset: data$D == 1)}

\item{data}{A data frame on which the linear model is estimated.}

\item{sel.bw}{SEL smoothing bandwidth.}

\item{eff.control}{NULL for the VS estimator, or a list passed to \code{rho.full.sample}
that may contain any of the following: \code{pi.hat}, \code{helper},
\code{helper.predicted}, \code{pi.bw}, \code{helper.bw}, \code{helper.degree}, \code{PIT}.}

\item{do.SE}{Logical: should the standard errors be computed numerically via
SEL Hessian inversion? Will be overriden with \code{TRUE} if \code{CI.lev}
is not \code{NULL} since the intervals rely on standard errors.}

\item{do.restr}{# Estimate models under constraints where one of the coefficients is equal to the true value?}

\item{CI.lev}{A numberic vector of probabilities for confidence intervale;
a good value is c(0.9, 0.95, 0.99). If \code{NULL}, the confidence intervals are not returned.}

\item{try.ellipse}{Logical: should the optimiser safeguard against potential
local optima (in very small samples) in some likely places based on the VCOV matrix?}

\item{...}{Passed to maximiseSEL}
}
\value{
A list with model type, SEL estimates, SEL-based variance-covariance
  matrix (if \code{do.SE}), constrained estimates and ELR test statistics
  (if \code{do.restr}), ELR-based confidence intervals (if CI.lev is not NULL), and smoothing bandwidths.
}
\description{
Estimation of one model via SEL
}
