% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{fdCoef}
\alias{fdCoef}
\title{Finite-difference coefficients for arbitrary grids}
\usage{
fdCoef(
  d = 1,
  order = 2,
  side = c("central", "forward", "backward"),
  stencil = NULL
)
}
\arguments{
\item{d}{Order of the derivative}

\item{order}{Order of accuracy in terms of the step size (usually
\eqn{h^o/o! \cdot f^{(o)}(x)}{h^o/o! * d^o/dx^o f(x)}).}

\item{side}{Centred or one-sided differences. Unless the function is computationally prohibitively expensive, two-sided differences are strongly recommended.}

\item{stencil}{In case the user desires a non-uniform grid or a custom grid, a vector of points at which the function is to be evaluated.}
}
\description{
This function computes the coefficients that yield an numerical approximation
or arbitrary order to the true first derivative of any function, or, for a given
grid (stencil) \emph{a}, the weights yielding the derivative approximation using the
function evaluated on the grid:
\deqn{\frac{df}{dx} \approx h^{-1} \sum_{i=1}^n a_i f(x + s_i\cdot h)}{f'(x) ~ sum_i a[i] f(x + s[i]*h)}
}
\details{
The finite-difference coefficients for any given stencil are given as a solution of a linear equation
system. There derivation of the system is due to \insertCite{taylor2016finite}{smoothemplik}, although a similar
approach is described in \insertCite{fornberg1988generation}{smoothemplik}. This function reproduces the tables
from the latter paper exactly.
}
\examples{
fdCoef() # Simple two-sided derivative
fdCoef(4)$weights # Should be (1/12, -2/3, 2/3, -1/12)
# Replicating Table 1 from Fornberg (1988) "Generation of Finite Difference Formulas..."
t1 <- t(sapply((1:4)*2, function(o) c(rep(0, 4-o/2), fdCoef(o)$weights, rep(0, 4-o/2))))
colnames(t1) <- (setdiff(-4:4, 0)); rownames(t1) <- (1:4)*2
print(t1, digits = 4)
# Replicating Table 2 (ibid) for a halfway stencil
s2 <- c(-7, -5, -3, -1, 1, 3, 5, 7)/2
t2 <- t(sapply(1:4, function(o) c(rep(0, 4-o), fdCoef(stencil = s2[(5-o):(o+4)])$weights,
                                  rep(0, 4-o))))
colnames(t2) <- s2; rownames(t2) <- (1:4)*2
print(t2, digits = 4)
# Replicating Table 3 (ibid)
t3 <- t(sapply(1:8, function(o) c(fdCoef(stencil = 0:o)$weights, rep(0, 8-o))))
colnames(t3) <- 0:8; rownames(t3) <- 1:8
print(t3, digits = 4)
}
\references{
\insertAllCited{}
}
