% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{g.unconditional.vs}
\alias{g.unconditional.vs}
\alias{g.unconditional.eff}
\title{Moment functions from the discrete design}
\usage{
g.unconditional.vs(theta, data)

g.unconditional.eff(theta, data, pihat, ystarhat)
}
\arguments{
\item{theta}{A parameter vector at which the moment function is evaluated.}

\item{data}{A data frame on which the moment function is evaluated.}

\item{pihat}{A numeric vector contatining the predicted propensity score. Must be the same length as \code{nrow(data)}.}

\item{ystarhat}{A numeric vector contatining the predicted Y* (notation from the paper). Must be the same length as \code{nrow(data)}.}
}
\value{
A matrix with two columns of the moment function values.
}
\description{
These functions are similar to \eqn{g(\theta, x)} described in the documentation of the \code{gmm} package.
}
\section{Functions}{
\itemize{
\item \code{g.unconditional.eff}: Moment function incorporating the propensity score.
}}

