% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{generateData}
\alias{generateData}
\title{Generation of data for Design 1 (continuous case)}
\usage{
generateData(
  n = 1000,
  distfun = stats::runif,
  eq.struc = function(z) 1 + 1 * z,
  eq.reduc = function(x) 1 + 1 * x,
  propensity = function(x) propensityScore(x),
  sigma2X = function(x) sigma2cragg(x),
  varUV = matrix(c(1, 1, 1, 2), ncol = 2),
  throwaway = 0,
  seed = 1
)
}
\arguments{
\item{n}{Number of observations.}

\item{distfun}{Density function of X}

\item{eq.struc}{A function computing the deterministic part of the structural equation}

\item{eq.reduc}{A function computing the deterministic part of the reduced form}

\item{propensity}{Propensity score function}

\item{sigma2X}{Skedastic function (conditional variance of U given X)}

\item{varUV}{Covariance matrix of U and V}

\item{throwaway}{Points on the right end to discard}

\item{seed}{Seed for Monte-Carlo simulation}
}
\value{
A data frame with variables necessary for Design 1 simulations.
}
\description{
Generation of data for Design 1 (continuous case)
}
