% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{generateDataDiscrete}
\alias{generateDataDiscrete}
\title{Generation of data for Design 2 (discrete case)}
\usage{
generateDataDiscrete(
  n = 1000,
  p = 0.3,
  params.struc = c(1, 1),
  params.reduc = c(0, 1),
  sigma2x = c(4, 1),
  varU = 1,
  varV = 2,
  covUV = 1,
  pix = c(0.3, 0.6),
  seed = 1
)
}
\arguments{
\item{n}{Number of observations.}

\item{p}{Probability of X=1 (Bernoulli)}

\item{params.struc}{Structural model parameter vector}

\item{params.reduc}{Reduced-form parameter vector}

\item{sigma2x}{Structural model error variance vector: Var(U | X) = x where x is 0 or 1}

\item{varU}{Structural model error variance multiplier}

\item{varV}{Reduced-form model error variance}

\item{covUV}{Covariance of the two errors, determining the degree of endogeneity}

\item{pix}{Propensity score evaluated at 0 and 1}

\item{seed}{Seed for Monte-Carlo simulation}
}
\value{
A data frame with variables necessary for Design 2 simulations.
}
\description{
Generation of data for Design 2 (discrete case)
}
