% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{getCoefELDiscrete}
\alias{getCoefELDiscrete}
\title{Estimate a Design-2 model with Owen's EL}
\usage{
getCoefELDiscrete(
  seed,
  design = list(n = 500, p = 0.6, params.struc = c(1, 1), params.reduc = c(0, 1),
    sigma2x = c(16, 1), pix = c(0.25, 0.9)),
  do.SE = TRUE,
  do.restr = TRUE,
  do.CI = TRUE
)
}
\arguments{
\item{seed}{The seed for data generation (passed to \code{generateDataDiscrete}).}

\item{design}{A list with design options (can have the following named elements: number of observations, structural parameters, reduced-form parameters, skedastic function values, propensity score values).}

\item{do.SE}{Logical: compute EL-based standard errors by inverting the numerically estimated Hessian and include results in the output?}

\item{do.restr}{Logical: Test the true null hypothesis that the coefficients are equal to their true values (one by one and jointly) via the ELR?}

\item{do.CI}{Logical: Compute the confidence regions for 90%, 95%, 97.5% and 99% levels?

Since the slope confidence regions are one-dimensional, unimodality of the ELR function is assumed, so the regions become intervals, possibly unbounded, and are found numerically)}
}
\value{
A list with the unrestricted estimates, constrained estimates and ELR test statistics (if \code{do.restr}),
EL confidence intervals (if \code{do.CI}), and EL-based standard errors (if \code{do.SE}).
}
\description{
Estimate a Design-2 model with Owen's EL
}
