% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{getCoefSELContinuous}
\alias{getCoefSELContinuous}
\title{Simulate and estimate one model with one seed and many hyper-parameters}
\usage{
getCoefSELContinuous(
  seed,
  design = list(),
  sel.bw,
  pi.bw = NULL,
  helper.bw = NULL,
  helper = NULL,
  degree = NULL,
  CV = FALSE,
  weak.instrument.f = 10,
  ...
)
}
\arguments{
\item{seed}{One numeric seed for Monte-Carlo simulations.}

\item{design}{A list containing the number of observations, skedastic function, and propensity score.}

\item{sel.bw}{A numeric vector of SEL smoothing bandwidths. Mandatory.}

\item{pi.bw}{NULL or a numeric vector of smoothing bandwidths for the propensity.}

\item{helper.bw}{NULL or a numeric vector of smoothing bandwidths for mu.}

\item{helper}{Type of intermediate object used in the prediction of mu.}

\item{degree}{Degree of non-parametric smoothing of the helper.}

\item{CV}{Logical: pick the smoothing bandwidth for pi(X) and the helper via cross-validation with PIT?}

\item{weak.instrument.f}{A scalar indicating the threshold F statistic: if the first-stage F is less than that, discard the simulation. Set to 0 to allow all simulations regardless of the instrument strength.}

\item{...}{Passed to estimateOneModelDesign1.}
}
\value{
A list with seed used, design parameters to recreate the data set,
  2SLS estimates, weak instrument diagnostics, optimal IV estimates, VS SEL
  estimates and, if proper smoothing parameters were passed, efficient SEL
  estimates, and time taken to compute the latter two.
}
\description{
Simulate and estimate one model with one seed and many hyper-parameters
}
