% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{kernelDensity}
\alias{kernelDensity}
\title{Kernel density estimation}
\usage{
kernelDensity(
  x,
  xgrid = NULL,
  bw = NULL,
  kernel = c("gaussian", "uniform", "triangular", "epanechnikov"),
  PIT = FALSE,
  return.grid = FALSE
)
}
\arguments{
\item{x}{A numeric vector or numeric matrix of predictors.}

\item{xgrid}{A numeric vector or numeric matrix with \code{ncol(xgrid) = ncol(x)} of points at which the density is estimated.}

\item{bw}{Bandwidth: a scalar or a vector of the same length as \code{ncol(x)}.}

\item{kernel}{Character describing the desired kernel type (Gaussian is infinitely smooth but does not provide finite support).}

\item{PIT}{Passed to \code{kernelWeights}.}

\item{return.grid}{Logical: if true, then returns \code{xgrid} and appends the estimated density as the last column.}
}
\value{
A vector of density estimates evaluated at the grid points or, if \code{return.gric}, a matrix with the density in the last column.
}
\description{
Kernel density estimation
}
