% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{kernelDiscreteDensitySmooth}
\alias{kernelDiscreteDensitySmooth}
\title{Density and/or kernel regression estimator with conditioning on discrete variables}
\usage{
kernelDiscreteDensitySmooth(x, y = NULL, compact = FALSE, fun = mean)
}
\arguments{
\item{x}{A vector or a matrix/data frame of discrete explanatory variables (exogenous). Non-integer values are fine because the data are split into bins defined by interactions of these variables.}

\item{y}{Optional: a vector of dependent variable values.}

\item{compact}{Logical: return unique values instead of full data with repeated observations?}

\item{fun}{A function that computes a statistic of \code{y} inside every category defined by \code{x}.}
}
\value{
A list with \code{x}, density estimator (\code{fhat}) and, if \code{y} was provided, regression estimate.
}
\description{
Density and/or kernel regression estimator with conditioning on discrete variables
}
