% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{lb.design1.theor}
\alias{lb.design1.theor}
\title{Calculation of efficiency bounds}
\usage{
lb.design1.theor(
  distfun = stats::dunif,
  support = c(0, 1),
  eq.reduc = function(x) 1 + 1 * x,
  propensity = function(x) propensityScore(x, lower = 0.75, upper = 0.25, r = 0.4, sd =
    0.05, spec = 5),
  sigma2X = function(x) sigma2cragg(x, r = -1/3, v = 2),
  varUV = matrix(c(1, 1, 1, 2), ncol = 2),
  rel.tol = 1e-06
)
}
\arguments{
\item{distfun}{The distribution function of the exogenous variable.}

\item{support}{The support over which \code{distfun} and its modifications should be integrated (passed to \code{integrate}).}

\item{eq.reduc}{The equation of the reduced form for the endogenous variable.}

\item{propensity}{The propensity score function.}

\item{sigma2X}{The skedastic function (conditional variance of errors).}

\item{varUV}{The covariance matrix of the baseline errors (U and V in our paper) of the structural and reduced equations.}

\item{rel.tol}{Desired relative accuracy of the numerical integration. Passed to \code{integrate}.}
}
\value{
A vector with 3 elements: the asymptotic variances of the inefficient slope estimator and the efficient slope estimator, and their ratio.
}
\description{
This function is to be used if the distribution of the conditioning variable (X in our paper) is continuous.
}
