% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{lb.design2.theor}
\alias{lb.design2.theor}
\title{Calculation of efficiency bounds}
\usage{
lb.design2.theor(
  p = 0.6,
  eq.reduc = function(x) 0 + 1 * x,
  propensity = c(0.25, 0.9),
  sigma2X = c(16, 1),
  varUV = matrix(c(1, 1, 1, 2), ncol = 2)
)
}
\arguments{
\item{p}{The expected value of X (the probability of being equal to 1 in the Bernoulli distribution).}

\item{eq.reduc}{The equation of the reduced form for the endogenous variable.}

\item{propensity}{The propensity score vector (\eqn{\pi(0, 1)}).}

\item{sigma2X}{The skedastic function vector (\eqn{\sigma^2(0, 1)}).}

\item{varUV}{The covariance matrix of the baseline errors (U and V in our paper) of the structural and reduced equations.}
}
\value{
A vector with 3 elements: the asymptotic variances of the inefficient slope estimator and the efficient slope estimator, and their ratio.
}
\description{
This function is to be used if the distribution of the conditioning variable (X in our paper) has a two-point support and no numerical integration is needed.
}
