% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{textWithHalo}
\alias{textWithHalo}
\alias{linesWithHalo}
\title{Adding text or lines with an outline for better legibility}
\usage{
textWithHalo(
  x,
  y,
  labels,
  nhalo = 32,
  col.halo = "#FFFFFF44",
  col.main = "#000000",
  hscale = 0.04,
  vscale = 0.04,
  ...
)

linesWithHalo(
  x,
  y,
  nhalo = 32,
  hscale = 0.01,
  vscale = 0.01,
  col.halo = "#FFFFFF88",
  col.main = "#000000",
  ...
)
}
\arguments{
\item{x}{Passed to the plotting command.}

\item{y}{Passed to the plotting command.}

\item{labels}{Passed to \code{text}.}

\item{nhalo}{The number of text duplicates to arrange in a circle to create the white glow effect.}

\item{col.halo}{Halo colour.}

\item{col.main}{Text colour.}

\item{hscale}{Halo horizontal radius.}

\item{vscale}{Halo vertical radius.}

\item{...}{Passed to the plotting command.}
}
\description{
Adding text or lines with an outline for better legibility
}
