% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfunctions.R
\name{lmEff}
\alias{lmEff}
\title{Newey's optimal IV estimator (1993) for linear models}
\usage{
lmEff(
  y,
  incl = NULL,
  endog = NULL,
  excl = NULL,
  bw = NULL,
  iterations = 2,
  coef.names = NULL
)
}
\arguments{
\item{y}{The vector containing dependent variable values}

\item{incl}{A vector or a matrix of included instruments (can be NULL)}

\item{endog}{A vector or a matrix of endogenous variables (can be NULL)}

\item{excl}{A vector or a matrix of excluded instruments (can be NULL)}

\item{bw}{A scalar of a vector or bandwidths to predict sigma2(incl, excl) (can be NULL)}

\item{iterations}{Iterations or Newey's weighting procedure; offers some finite-sample (but not asymptotic) improvement if greater than 1}

\item{coef.names}{A character vector to assign as the names of the estimated coefficients (useful if some variables are endogenous and their projections inherit uninformative names)}
}
\value{
A list with optimal IV estimates, their variance-covariance matrix, and model residuals.
}
\description{
Newey's optimal IV estimator (1993) for linear models
}
\examples{
d <- generateData(seed = 1, n = 500)
d <- d[as.logical(d$D), ]
incl <- NULL
endog <- d$Z
excl <- d$X
y <- d$Y
lmEff(y, incl, endog, excl)
m1 <- AER::ivreg(y ~ endog | excl)
m2 <- lmEff(y, incl, endog, excl, iterations = 0)
}
