% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scelcount.R
\name{logTaylor}
\alias{logTaylor}
\title{r-th derivative of the kth-order Taylor expansion of log(x)}
\usage{
logTaylor(x, a = 1, k = 4, r = 0)
}
\arguments{
\item{x}{Numeric: a vector of points for which the logarithm is to be evaluated}

\item{a}{Scalar: the point at which the polynomial approximation is computed}

\item{k}{Positive integer: maximum polynomial order in the Taylor expansion of the original function.}

\item{r}{Non-negative integer: derivative order

Note that this function returns the r-th derivative of the k-th-order Taylor expansion, not the
k-th-order approximation of the r-th derivative. Therefore, the degree of the resulting polynomial
is \eqn{r-k}.}
}
\value{
The approximating Taylor polynomial around \code{a} of the order \code{r-k} evaluated at \code{x}.
}
\description{
r-th derivative of the kth-order Taylor expansion of log(x)
}
\examples{
cl <- c("#e6194B", "#3cb44b", "#ffe119", "#4363d8", "#f58231",
        "#911eb4", "#42d4f4", "#f032e6", "#bfef45")
a <- 1.5
x <- seq(a*2, a/2, length.out = 101)
f <- function(x, r = 0)  if (r == 0) log(x) else ((r\%\%2 == 1)*2-1) * 1/x^r * gamma(r)
par(mfrow = c(2, 3), mar = c(2, 2, 2.5, 0.2))
for (r in 0:5) {
y <- f(x, r = r)
plot(x, y, type = "l", lwd = 7, bty = "n", ylim = range(0, y),
       main = paste0("d^", r, "/dx^", r, " Taylor(Log(x))"))
  for (k in 0:8) lines(x, logTaylor(x, a = a, k = k, r = r), col = cl[k+1], lwd = 1)
  points(a, f(a, r = r), pch = 16, cex = 1.5, col = "white")
}
legend("topright", as.character(0:8), title = "Order", col = cl, lwd = 1)
}
