% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{LSCV}
\alias{LSCV}
\title{Least-squares cross-validation function for the Nadaraya-Watson estimator}
\usage{
LSCV(
  x,
  y,
  bw,
  same = FALSE,
  degree = 0,
  kernel = "gaussian",
  order = 2,
  robust.iterations = 0
)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{y}{A vector of dependent variable values.}

\item{bw}{Candidate bandwidth values: scalar, vector, or a matrix (with columns corresponding to columns of \code{x}).}

\item{same}{Logical: use the same bandwidth for all columns of \code{x}?}

\item{degree}{Passed to \code{kernelSmooth}.}

\item{kernel}{Passed to \code{kernelSmooth}.}

\item{order}{Passed to \code{kernelSmooth}.}

\item{robust.iterations}{Passed to \code{kernelSmooth}.}
}
\value{
A numeric vector of the same length as \code{bw} or \code{nrow(bw)}.
}
\description{
Least-squares cross-validation function for the Nadaraya-Watson estimator
}
