% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{matrix2LaTeX}
\alias{matrix2LaTeX}
\title{Converting matrices to LaTeX tables}
\usage{
matrix2LaTeX(x, d = 3)
}
\arguments{
\item{x}{A number vector or matrix}

\item{d}{Passed to trail0.}
}
\value{
No return, just prints the output.
}
\description{
Converting matrices to LaTeX tables
}
