% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfunctions.R
\name{maximiseSEL}
\alias{maximiseSEL}
\title{Title}
\usage{
maximiseSEL(
  rho,
  start.values = NULL,
  restricted.params = NULL,
  sel.weights = NULL,
  verbose = FALSE,
  parallel = FALSE,
  cores = 1,
  optmethod = c("nlm", "BFGS", "Nelder-Mead"),
  nlm.step.max = NULL,
  maxit = 50,
  ...
)
}
\arguments{
\item{rho}{Passed to \code{\link{smoothEmplik}} and \code{\link{constrSmoothEmplik}}.}

\item{start.values}{Initial values for unrestricted parameters.}

\item{restricted.params}{Fixed parameter values used for constrained optimisation, hypothesis testing, power and size calculations etc.}

\item{sel.weights}{Passed to \code{\link{smoothEmplik}} and \code{\link{constrSmoothEmplik}}}

\item{verbose}{If TRUE, reports optimisation progress, otherwise remains silent.}

\item{parallel}{Passed to \code{\link{smoothEmplik}} and \code{\link{constrSmoothEmplik}}.}

\item{cores}{Passed to \code{\link{smoothEmplik}} and \code{\link{constrSmoothEmplik}}.}

\item{optmethod}{A string indicating the optimisation method: "nlm" (the default, invoking \code{stats::nlm}) is slightly quicker, and if it fails, "BFGS" is used}

\item{nlm.step.max}{Passed to \code{nlm} if \code{optmethod == "nlm"}; in case of convergence issues, can be reduced, but if it is too small and 5 \code{nlm} iterations are done with the maximum step size, optimisation is re-done via BFGS.}

\item{maxit}{Maximum number of numerical optimiser steps. If it has not converged in this number of steps, fail gracefully with a meaningfull return.}

\item{...}{Passed to \code{\link{smoothEmplik}} and \code{\link{constrSmoothEmplik}}.}
}
\value{
A list with the following elements:
\itemize{
\item{par }{The argmax of the SEL that was found by the optimiser.}
\item{value }{The SEL value corresponding to \code{par}.}
\item{restricted }{A logical vector of the same length as \code{par} indicating if the respective element was kept fixed during the optimisation.}
\item{xtimes }{A vector with two components indicating time (in seconds) it took to find the candidate initial value and the final optimum respectively.}
}
}
\description{
Title
}
