% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scelcount.R
\name{mllog}
\alias{mllog}
\title{Modified minus logarithm with derivatives}
\usage{
mllog(x, eps = NULL, M = Inf, der = 0, order = 4, flatten1d = TRUE)
}
\arguments{
\item{x}{Numeric vector for which approximated logarithm is to be computed.}

\item{eps}{Lower threshold below which approximation starts.}

\item{M}{Upper threshold above which approximation starts.}

\item{der}{Non-negative integer: 0 yields the function, 1 and higher yields derivatives}

\item{order}{Positive integer: Taylor approximation order.}

\item{flatten1d}{If TRUE and \code{x} is a vector, the output is a vector and not a 1-column matrix.}
}
\value{
A numeric matrix with \code{(order+1)} columns containing the values of the modified log and its derivatives.
}
\description{
Modified minus logarithm with derivatives
}
\details{
Provides a family of alternatives to -log() and derivative thereof in order to attain self-concordance and
computes the modified negative logarithm and its first derivatives.
For eps <= x <= M, returns just the logarithm. For x < eps and x > M, returns the Taylor approximation of the given \code{order}.
4th order is the lowest that gives self concordance.
}
\examples{
x <- seq(0.01^0.25, 2^0.25, length.out = 51)^4 - 0.11 # Denser where |f'| is higher
plot(x, -log(x))
lines(x, mllog(x, eps = 0.2), col = 2)
lines(x, mllog(x, eps = 0.5), col = 3)
lines(x, mllog(x, eps = 1, M = 1.2, order = 6), col = 4)
# Suppose that we substitute -log with its Taylor approx. around 1
x <- seq(0.1, 1.5, 0.2)
err.mat <- cbind(x, ErrOrder2 = -log(x) - mllog(x, eps = 1, M = 1, order = 2),
                    ErrOrder3 = -log(x) - mllog(x, eps = 1, M = 1, order = 3),
                    ErrOrder4 = -log(x) - mllog(x, eps = 1, M = 1, order = 4),
                    ErrOrder5 = -log(x) - mllog(x, eps = 1, M = 1, order = 5),
                    ErrOrder6 = -log(x) - mllog(x, eps = 1, M = 1, order = 6))
print(err.mat, digits = 2)
}
