% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{propensityScore}
\alias{propensityScore}
\alias{propensityScore.discrete}
\alias{sigma2cragg}
\alias{sigma2.discrete}
\title{Design functions for simulation}
\usage{
propensityScore(x, spec = 5, r = 0.1, lower = 0.95, upper = 0.25, sd = 0.5)

propensityScore.discrete(x, s)

sigma2cragg(x, r = -1/3, v = 2)

sigma2.discrete(x, s)
}
\arguments{
\item{x}{A numeric vector for which the propensity function is evaluated.}

\item{spec}{An integer indicating the propensity specification type, an integer from 1 to 5.}

\item{r}{A numeric scalar of the location parameter (the shift of the propensity or skedastic function).}

\item{lower}{A numeric scalar, the lower bound for the sigmoid (specification 5).}

\item{upper}{A numeric scalar, the upper bound for the sigmoid (specification 5).}

\item{sd}{A numeric scalar, the standard deviation of the normal distribution that yields the sigmoid for specification 5 (i.e. sigmoid smoothness).}

\item{s}{The values of the propensity score function at \eqn{X = 0} and \eqn{X = 1} (discrete case).}

\item{v}{The power to which the absolute value of the error is raised in the skedastic function.}
}
\value{
A numeric vector of the same length as \code{x} with the propensity score function values (between 0 and 1).
}
\description{
The propensity score functions from Wang and Rao with possibly one of three specifications, as well as two special cases (always 1 or sigmoid).
}
\section{Functions}{
\itemize{
\item \code{propensityScore.discrete()}: The propensity score for the discrete case.

\item \code{sigma2cragg()}: The skedastic function for the continuous case.

\item \code{sigma2.discrete()}: The skedastic function for the discrete case.

}}
