% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{rho.complete.case}
\alias{rho.complete.case}
\alias{rho.complete.case.pi}
\title{Moment functions with conditioning}
\usage{
rho.complete.case(theta, data)

rho.complete.case.pi(theta, data, pi.hat)
}
\arguments{
\item{theta}{A parameter vector at which the moment function is evaluated.}

\item{data}{A data frame on which the moment function is evaluated.

Unlike \code{g.unconditional.vs} and \code{g.unconditional.eff}, these functions are designed for models with conditional moment restrictions.}

\item{pi.hat}{A numeric vector contatining the predicted propensity score. Must be the same length as \code{nrow(data)}.}
}
\value{
A matrix with two columns of the moment function values.
}
\description{
Moment functions with conditioning
}
\section{Functions}{
\itemize{
\item \code{rho.complete.case.pi}: Moment function depending on the propensity score.
}}

