% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfunctions.R
\name{smoothEmplikList}
\alias{smoothEmplikList}
\title{SEL for every observation}
\usage{
smoothEmplikList(
  rho,
  sel.weights = NULL,
  parallel = FALSE,
  cores = 1,
  grad = FALSE,
  ...
)
}
\arguments{
\item{rho}{The moment function depending on parameters and data (and potentially other parameters). Must return a numeric vector.}

\item{sel.weights}{A matrix with valid kernel smoothing weights with rows adding up to 1, or a list of kernel weights for smoothing where the sum of each element is 1 (must be returned by \code{sparseVectorToList}).}

\item{parallel}{If TRUE, uses \code{parallel::mclapply} to speed up the computation.}

\item{cores}{The number of cores used by \code{parallel::mclapply}.}

\item{grad}{Currently not used.}

\item{...}{Passed to \code{rho}; usually \code{theta} and \code{data}.}
}
\value{
A list with two values:
\item{rho}{One-dimensional vector of the moment function values.}
\item{empliklist}{A list with the results of one-dimension optimisation conducted by \code{\link{weightedEL}}.}
}
\description{
SEL for every observation
}
