% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scel.R
\name{tracelr}
\alias{tracelr}
\title{Compute empirical likelihood on a trajectory}
\usage{
tracelr(z, mu0, mu1, N, verbose = FALSE, ...)
}
\arguments{
\item{z}{Passed to \code{emplik}.}

\item{mu0}{Starting point of trajectory}

\item{mu1}{End point of trajectory}

\item{N}{Number of segments into which the path is split (i. e. \code{N+1} steps are used).}

\item{verbose}{Logical: report iteration data?}

\item{...}{Passed to \code{emplik}.}
}
\value{
A matrix with one row at each mean from mu0 to mu1 and a column for each EL return value (except EL weights).
}
\description{
Compute empirical likelihood on a trajectory
}
