% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smdfunctions.R
\name{weightedT}
\alias{weightedT}
\title{Weighted t-statistic}
\usage{
weightedT(z, mu = 0, ct.mean, ct.var)
}
\arguments{
\item{z}{Numeric data vector.}

\item{mu}{Hypothesized mean of \code{z} in the moment condition.}

\item{ct.mean}{Numeric vector of weights for the variance (must sum to something > 0).}

\item{ct.var}{Numeric vector of weights for the mean (must sum to something > 0).}
}
\value{
A numeric scalar: \eqn{t = (\bar{y}_w-\mu) / \sqrt{\sigma_w^2}}{t = wmean/sqrt(wvar)}.
}
\description{
Computes a weighted t-statistic for a numeric vector \code{y} to test the hypothesis that the
weighted mean of \code{y} is \code{mu}, with possibly different weights for mean and varianace
estimations.  The statistic is the weighted mean of \code{y - mu} divided by
the square root of its weighted variance around that mean.
}
\details{
Note that this t statistic does *not* take into account the sample size or the number of
observations in the vicinity; this quantity does not adapt to the bandwidth size
}
\examples{
x <- 1:6
w <- 6:1
weightedT(x, mu = 3.5, w, w)
}
