% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{kernelDensity}
\alias{kernelDensity}
\title{Kernel density estimation}
\usage{
kernelDensity(
  x,
  xgrid = NULL,
  bw = NULL,
  gaussian = TRUE,
  method = "CPP",
  return.grid = FALSE
)
}
\arguments{
\item{x}{A numeric vector or numeric matrix of predictors.}

\item{xgrid}{A numeric vector or numeric matrix with \code{ncol(xgrid) = ncol(x)} of points at which the density is estimated.}

\item{bw}{Bandwidth: a scalar or a vector of the same length as \code{ncol(x)}.}

\item{gaussian}{If TRUE, the normal (Gaussian) product kernel with full support is used, otherwise Epanechnikov.}

\item{method}{"R" or "CPP". If "CPP", then faster Rcpp functions are invoked, otherwise use a pure R implementation.}

\item{return.grid}{Logical: if true, then returns \code{xgrid} and appends the estimated density as the last column.}
}
\value{
A vector of density estimates evaluated at the grid points or, if \code{return.gric}, a matrix with the density in the last column.
}
\description{
Kernel density estimation
}
