% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{kernelMixedDensity}
\alias{kernelMixedDensity}
\title{Density with conditioning on discrete and continuous variables}
\usage{
kernelMixedDensity(x, by, xgrid = NULL, ...)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{by}{A variable containing unique identifiers of discrete categories.}

\item{xgrid}{A numeric vector or a numeric matrix.}

\item{...}{Passed to \code{kernelDensity}.}
}
\value{
A numeric vector of the density estimate of the same length as \code{nrow(xgrid)}.
}
\description{
Density with conditioning on discrete and continuous variables
}
