% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothers.R
\name{kernelSmooth}
\alias{kernelSmooth}
\title{Local kernel smoother}
\usage{
kernelSmooth(
  x,
  y,
  xgrid = NULL,
  bw = NULL,
  gaussian = TRUE,
  LOO = FALSE,
  method = "CPP",
  degree = 0,
  standardise = TRUE,
  trim = function(x) 0.01/length(x),
  robust.iterations = 0,
  return.grid = FALSE
)
}
\arguments{
\item{x}{A vector or a matrix of exogenous explanatory variables.}

\item{y}{A vector of dependent variable values.}

\item{xgrid}{A vector or a matrix with the same number of columns as \code{x} contatining the values at which to evaluate the smoother.}

\item{bw}{A scalar or a vector of smoothing bandwidths.}

\item{gaussian}{If \code{TRUE}, Gaussian kernel is used, otherwise Epanechnikov.}

\item{LOO}{If \code{TRUE}, the leave-one-out estimator is returned.}

\item{method}{\code{"R"} or \code{"CPP"}. The latter requires \code{Rcpp} compilation.}

\item{degree}{Integer: 0 for locally constant estimator (Nadaraya---Watson), 1 for locally linear (Cleveland's LOESS), 2 for locally quadratic (use with care)}

\item{standardise}{If \code{TRUE}, then the variables are centred to have mean 0 and re-scaled to have standard deviation 1.}

\item{trim}{Trimming function for small weights to speed up local regression (if \code{degree} is 1 or 2).}

\item{robust.iterations}{The number of robustifying iterations (due to Cleveland, 1979).}

\item{return.grid}{If \code{TRUE}, prepends \code{xgrid} to the return results.

Standardisation is recommended for the purposes of numerical stability (sometimes \code{lm()} might choke when the dependent variable takes very large absolute values and its square is used).}
}
\value{
A vector of predicted values or, if \code{return.grid} is \code{TRUE}, a matrix with the predicted values in the last column.
}
\description{
Local kernel smoother
}
