% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scel.R
\name{mllog}
\alias{mllog}
\title{Modified minus logarithm with derivatives}
\usage{
mllog(x, eps, M = Inf, der = 0)
}
\arguments{
\item{x}{Numeric vector for which approximated logarithm is to be computed.}

\item{eps}{Lower threshold below which approximation starts.}

\item{M}{Upper threshold above which approximation starts.}

\item{der}{Integer: 0, 1, or 2 derivatives}
}
\value{
A numeric matrix with 1, 2, or 3 columns containing the modified log and possibly its two first derivatives.
}
\description{
Modified minus logarithm with derivatives
}
\details{
Provides a family of alternatives to -log() and derivative thereof in order to attain self-concordance and
computes the modified negative logarithm and its first derivatives.
For eps <= x <= M, returns just the logarithm. For x < eps and x > M, returns the 4th-order Taylor approximation.
4th order is the lowest that gives self concordance.
}
