% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfunctions.R
\name{smoothEmplik}
\alias{smoothEmplik}
\title{Smoothed Empirical Likelihood function value}
\usage{
smoothEmplik(
  rho,
  sel.weights = NULL,
  trim = NULL,
  minus = FALSE,
  parallel = FALSE,
  cores = 1,
  bad.value = -Inf,
  ...
)
}
\arguments{
\item{rho}{Passed to \code{\link{smoothEmplikList}}.}

\item{sel.weights}{Passed to \code{\link{smoothEmplikList}}.}

\item{trim}{A vector of trimming function values to multiply the output of \code{rho(...)} with. If NULL, no trimming is done.}

\item{minus}{If TRUE, returns SEL times -1 (for optimisation via minimisation).}

\item{parallel}{Passed to \code{\link{smoothEmplikList}}.}

\item{cores}{Passed to \code{\link{smoothEmplikList}}.}

\item{bad.value}{Replace non-finite individual SEL values with this value. May be useful if the optimiser does not allow specific non-finite values.}

\item{...}{Passed to \code{\link{smoothEmplikList}}.}
}
\value{
A scalar with the SEL value.
}
\description{
Evaluates SEL function for a given moment function at a certain parameter value.
}
