% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scel.R
\name{svdlm}
\alias{svdlm}
\title{Least-squares regression via SVD}
\usage{
svdlm(X, y)
}
\arguments{
\item{X}{Model matrix.}

\item{y}{Response vector.

Empirical likelihood's Newton steps are of least-squares type.
Denote \eqn{X^+} to be the generalised inverse of X. If SVD algorithm failures are encountered,
it sometimes helps to try \code{svd(t(X))} and translate back. First check to ensure that X does not contain \code{NaN}, or \code{Inf}, or \code{-Inf}.}
}
\value{
A matrix of coefficients.
}
\description{
Least-squares regression via SVD
}
