% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfunctions.R
\name{weightedEL}
\alias{weightedEL}
\title{Empirical likelihood for one-dimensional vectors}
\usage{
weightedEL(
  z,
  ct = NULL,
  shift = NULL,
  mu = 0,
  SEL = FALSE,
  n.orig = NULL,
  weight.tolerance = 0.01/length(z),
  truncto = 0,
  uniroot.control = list(),
  return.weights = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{z}{The data vector.}

\item{ct}{The count variable that indicates the multiplicity ob observations. Can be fractional. Very small counts below the tolerance threshold are zeroed.}

\item{shift}{The value to add in the denominator (useful in case there are extra Lagrange multipliers): 1 + lambda'Z + shift.}

\item{mu}{Hypothesized mean of \code{z} in the moment condition.}

\item{SEL}{If FALSE, then the boundaries for the lambda search are based on the total sum of counts, like in vanilla empirical likelihood,
due to Owen (2001, formula 2.9), otherwise according to Cosma et al. (2019, p. 170, the topmost formula).}

\item{n.orig}{An optional scalar to denote the original sample size (useful in the rare cases re-normalisation is needed).}

\item{weight.tolerance}{Weight tolerance for counts for numerical stability (similar to the ones in Art B. Owen's 2017 code, but adapting to the sample size).}

\item{truncto}{Counts under \code{weight.tolerance} will be set to this value. In most cases, setting this to \code{0} or \code{weight.tolerance} is a viable solution of the zero-denominator problem.}

\item{uniroot.control}{A list passed to the \code{uniroot}.}

\item{return.weights}{Logical: if TRUE, individual EL weights are computed and returned. Setting this to FALSE gives huge memory savings in large data sets, especially when smoothing is used.}

\item{verbose}{Logical: if \code{TRUE}, prints warnings.}
}
\value{
A list with the following elements:
\itemize{
\item{logelr }{Logarithm of the empirical likelihood ratio.}
\item{lam }{The Lagrange multiplier.}
\item{wts }{Observation weights/probabilities (of the same length as \code{z}).}
\item{converged }{TRUE if the algorithm converged, FALSE otherwise (usually means that \code{mu} is not in the convex hull of the data, that is, within the range of \code{z}).}
\item{iter }{The number of iterations used (from \code{uniroot}).}
\item{bracket }{The admissible interval for lambda (that is, yielding weights between 0 and 1).}
\item{estim.prec }{The approximate estimated precision of lambda (from \code{uniroot}).}
\item{f.root }{The value of the function at the root (from \code{uniroot}); values \code{> sqrt(.Machine$double.eps)} indicate covergence problems.}
\item{exitcode }{An integer indicating the reason of termination.}
\describe{
\item{0:}{success, an interior solution for lambda found.}
\item{1:}{the value of the derivative is \code{> sqrt(.Machine$double.eps)} (the value of the function at the returned root it not exactly zero, and this is not an issue with the tolerance).}
\item{2:}{the root was found and the function value seems to be zero, but the root is very close (\code{< sqrt(.Machine$double.eps)}) to the boundary.}
\item{3:}{like \code{1} and \code{2} simultaneously: the value of the target function is > 1e-8, and the result is close to the boundary (< 1e-8).}
\item{4:}{an error occurred while calling \code{uniroot}.}
\item{5:}{\code{mu} is not strictly in the convex hull of \code{z} (spanning condition not met).}
}
}
}
\description{
Empirical likelihood with counts to solve one-dimensional problems efficiently with Brent's root search algorithm. Conducts an empirical likelihood ratio test of the hypothesis that the mean of \code{z} is \code{mu}
The names of the elements in the returned list are consistent with Art B. Owen's original R code.
}
\references{
Owen, A. B. (2001). *Empirical likelihood*. CRC press.

Cosma, A., Kostyrka, A. V., & Tripathi, G. (2019). Inference in conditional moment restriction models when there is selection due to stratification. In *The Econometrics of Complex Survey Data*. Emerald Publishing Limited.
}
